edit storyboard
